Remove noOfAccessPointsSuccessfullyProcessedWithWarnings field from output document
We are removing the (optional) field noOfAccessPointsSuccessfullyProcessedWithWarnings from the OutputDocument, since it was misleading. The number of stores with warnings will be included in the noOfAccessPointsSuccessfullyProcessed, since warnings don't prevent the stores from being processed. We are keeping the field warningDetails, which provide details which stores have warnings for your visibility.